src/requirements.txt
python-dateutil==2.7.3
boto3==1.7.57
botocore==1.10.57
